+++ weblate/trans/models/workflow.py
@@ -4,6 +4,7 @@
 
 from __future__ import annotations
 
+from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.db import models
 from django.utils.translation import gettext, gettext_lazy
@@ -21,7 +22,7 @@ class WorkflowSetting(models.Model):
     # This should match definitions in Project
     translation_review = models.BooleanField(
         verbose_name=gettext_lazy("Enable reviews"),
-        default=False,
+        default=settings.DEFAULT_TRANSLATION_REVIEW,
         help_text=gettext_lazy("Requires dedicated reviewers to approve translations."),
     )
     # This should match definition in Component