+++ weblate/vcs/mercurial.py
@@ -400,7 +400,11 @@ class HgRepository(Repository):
             ["cat", "--rev", revision, path], needs_lock=False, merge_err=False
         )
 
-    def cleanup(self) -> None:
+    def remove_stale_branches(self) -> None:
+        """Remove stale branches and tags from the repository."""
+        return
+
+    def cleanup_files(self) -> None:
         """Remove not tracked files from the repository."""
         self.set_config_values(("extensions", "purge", ""))
         self.execute(["purge"])